requirements.txt
astroid==2.4.2
autopep8==1.5.4
cachetools==4.1.1
certifi==2020.12.5
chardet==3.0.4
flake8==3.8.4
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-pubsub==2.2.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.34.0
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
libcst==0.3.15
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.0
proto-plus==1.11.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.6.0
PySocks==1.7.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
toml==0.10.2
tweepy==3.9.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
wrapt==1.12.1
